Qgis3.0 migration

+++ gui/DatabaseTools/UserTools/serverProfilesManager.py
@@ -33,8 +33,8 @@ from qgis.PyQt.QtGui import QCursor
 from qgis.core import QgsMessageLog
 
 # DSGTools imports
-from DsgTools.UserTools.createProfileWithProfileManager import CreateProfileWithProfileManager
-from DsgTools.CustomWidgets.genericParameterSetter import GenericParameterSetter
+from DsgTools.gui.DatabaseTools.UserTools.createProfileWithProfileManager import CreateProfileWithProfileManager
+from DsgTools.gui.CustomWidgets.DatabasePropertiesWidgets.BasicPropertyWidgets.genericParameterSetter import GenericParameterSetter
 
 import json
 
@@ -61,7 +61,7 @@ class ServerProfilesManager(QtWidgets.QDialog, FORM_CLASS):
         """
         Creates tree widget items
         """
-        item = QtGui.QTreeWidgetItem(parent)
+        item = QtWidgets.QTreeWidgetItem(parent)
         item.setFlags(QtCore.Qt.ItemIsEnabled | QtCore.Qt.ItemIsTristate | QtCore.Qt.ItemIsUserCheckable)
         item.setCheckState(1, QtCore.Qt.Unchecked)
         item.setCheckState(2, QtCore.Qt.Unchecked)
@@ -217,19 +217,20 @@ class ServerProfilesManager(QtWidgets.QDialog, FORM_CLASS):
     
     @pyqtSlot(bool)
     def on_deletePermissionPushButton_clicked(self):
-        profileName, edgvVersion = self.profilesListWidget.currentItem().text().split(' (')
-        edgvVersion = edgvVersion.replace(')','')
-        if QMessageBox.question(self, self.tr('Question'), self.tr('Do you really want to delete profile ')+profileName+'?', QMessageBox.Ok|QMessageBox.Cancel) == QMessageBox.Cancel:
-            return
-        try:
-            QApplication.setOverrideCursor(QCursor(Qt.WaitCursor))
-            self.permissionManager.deleteSetting(profileName, edgvVersion)
-            QApplication.restoreOverrideCursor()
-            QMessageBox.warning(self, self.tr('Success!'), self.tr('Permission ') + profileName + self.tr(' successfully deleted.'))
-            self.refreshProfileList()
-        except Exception as e:
-            QApplication.restoreOverrideCursor()
-            QMessageBox.warning(self, self.tr('Warning!'), self.tr('Error! Problem deleting permission: ') + ':'.join(e.args))
+        if self.profilesListWidget.currentItem() is not None:
+            profileName, edgvVersion = self.profilesListWidget.currentItem().text().split(' (')
+            edgvVersion = edgvVersion.replace(')','')
+            if QMessageBox.question(self, self.tr('Question'), self.tr('Do you really want to delete profile ')+profileName+'?', QMessageBox.Ok|QMessageBox.Cancel) == QMessageBox.Cancel:
+                return
+            try:
+                QApplication.setOverrideCursor(QCursor(Qt.WaitCursor))
+                self.permissionManager.deleteSetting(profileName, edgvVersion)
+                QApplication.restoreOverrideCursor()
+                QMessageBox.warning(self, self.tr('Success!'), self.tr('Permission ') + profileName + self.tr(' successfully deleted.'))
+                self.refreshProfileList()
+            except Exception as e:
+                QApplication.restoreOverrideCursor()
+                QMessageBox.warning(self, self.tr('Warning!'), self.tr('Error! Problem deleting permission: ') + ':'.join(e.args))
     
     @pyqtSlot(bool)
     def on_saveButton_clicked(self):
